refactor: fileContact
