+++ src/lib/contact/contact.ts
@@ -13,13 +13,16 @@ import type {
 } from 'types/types';
 
 import { ALEPH_CHANNEL } from 'config/constants';
+import ContactFile from './contactClasses/fileContact';
 
 class Contact {
 	public contacts: IPCContact[];
 
 	public username: string;
 
-	private readonly account: accounts.ethereum.ETHAccount;
+	public account: accounts.ethereum.ETHAccount;
+
+	public files: ContactFile;
 
 	constructor(importedAccount: accounts.ethereum.ETHAccount) {
 		this.contacts = [];
@@ -27,7 +30,7 @@ class Contact {
 		this.username = '';
 	}
 
-	private async publishAggregate(): Promise<AggregateMessage<AggregateContentType>> {
+	public async publishAggregate(): Promise<AggregateMessage<AggregateContentType>> {
 		const aggr = await aggregate.Get<AggregateType>({
 			address: this.account.address,
 			keys: ['InterPlanetaryCloud'],
@@ -255,43 +258,6 @@ class Contact {
 		} catch (err) {
 			console.error(err);
 			return { success: false, message: 'Failed to update the file content' };
-		}
-	}
-
-	public async updateFileName(concernedFile: IPCFile, newName: string, sharedFiles: IPCFile[]): Promise<ResponseType> {
-		try {
-			let fileFound = false;
-			await Promise.all(
-				this.contacts.map(async (contact) => {
-					const file = contact.files.find((f) => f.id === concernedFile.id);
-					if (file) {
-						file.name = newName;
-						file.logs.push({
-							action: `Renamed file to ${newName}`,
-							date: Date.now(),
-						});
-						fileFound = true;
-						await this.publishAggregate();
-					}
-				}),
-			);
-			if (!fileFound) {
-				const file = sharedFiles.find((f) => f.id === concernedFile.id);
-				if (!file) {
-					return { success: false, message: 'File not found' };
-				}
-				await post.Publish({
-					account: this.account,
-					postType: 'InterPlanetaryCloud',
-					content: { file: { ...concernedFile, name: newName }, tags: ['rename', concernedFile.id] },
-					channel: 'TEST',
-					storageEngine: ItemType.ipfs,
-				});
-			}
-			return { success: true, message: 'Filename updated' };
-		} catch (err) {
-			console.error(err);
-			return { success: false, message: 'Failed to update this filename' };
 		}
 	}
