+++ src/lib/user.ts
@@ -2,7 +2,7 @@ import { accounts } from 'aleph-sdk-ts';
 import { aggregate } from 'aleph-sdk-ts/dist/messages';
 
 import Computing from 'lib/computing';
-import Contact from 'lib/contact';
+import Contact from 'lib/contact/contact';
 import Drive from 'lib/drive';
 
 import { AggregateType, IPCConfig, IPCContact } from 'types/types';
